+++ web/src/game/scenes/GameOver.ts
@@ -1,4 +1,4 @@
-import { EventBus } from '../EventBus';
+import { EventQueue } from '../event-queue';
 import { Scene } from 'phaser';
 
 export class GameOver extends Scene {
@@ -29,7 +29,7 @@ export class GameOver extends Scene {
             .setOrigin(0.5)
             .setDepth(100);
 
-        EventBus.emit('current-scene-ready', this);
+        EventQueue.emit('current-scene-ready', this);
     }
 
     changeScene() {