+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # CHANGELOG
 
-## 0.1.0 - 2024-07-17
+## 0.1.0 - 2025-06-05
 ### Added:
 - 新增翻譯模組，支援 Pipe、Directive、Service 三種形式。字典放在 `public/locale/` 下。
+- 導入 phaser 網頁遊戲引擎，並成功搭建 Angular 網頁遊戲基本框架。
+- 導入 CodeQL 作為 Code Review。
 
 ## first commit - 2024-07-16